+++ src/content/experience/index.ts
@@ -1,4 +1,7 @@
 // src/content/experience/index.ts
 // barrel exports for experience content
 
-export { WORK_EXPERIENCE } from '~/content/experience/workExperience'
+export {
+  EXPERIENCE_CONTENT,
+  WORK_EXPERIENCE,
+} from '~/content/experience/workExperience'